+++ src/game/game.js
@@ -1,9 +1,13 @@
+const generator = require('./../../generator');
+
 class Game {
 
     constructor() {
         this.startable = false;
         this.tree = [];
         this.demoUser = null;
+        this.villages = [];
+        this.fields = [];
     }
 
     canStart() {
@@ -44,6 +48,31 @@ class Game {
 
     addDemoUser(user) {
         this.demoUser = user;
+    }
+
+    addVillage(owner, village) {
+        let yid = generator.generateID();
+        this.villages.push({ yid: yid, owner: owner, village: village });
+        // generate fields
+        let resources = this.tree.buildings[0].building.res;
+        for (let i = 0; i < resources.length; i++) {
+            for (let i = 0; i < 3; i++) {
+                let field = {
+                    yid: generator.generateID(),
+                    village: yid,
+                    name: resources[i].name,
+                };
+                this.fields.push(field);
+            }
+        }
+    }
+
+    numberOfVillages() {
+        return this.villages.length;
+    }
+
+    numberOfFields() {
+        return this.fields.length;
     }
 
 }