+++ src/game/game.test.js
@@ -2,6 +2,7 @@ const Game = require('./game')
 const Tree = require('./../tree/tree')
 const Building = require('./../building/building')
 const Unit = require('./../unit/unit')
+const Village = require('./../village/village')
 
 const game = new Game();
 
@@ -52,4 +53,25 @@ test('game cant start without first demo user', () => {
     });
 
     expect(game.canStart()).toBe(true);
+})
+
+test('game add village after user creation', () => {
+    let tree = new Tree();
+    let castle = new Building();
+    let firstUnit = new Unit();
+
+    tree.addBuilding('castle', castle);
+    game.addBuildingTreeAndUnits(tree, [firstUnit]);
+    game.grandUnitBuildiner({
+        building: 'castle',
+        level: 2,
+    });
+    game.addDemoUser({
+        username: 'sensorario',
+        password: 'password',
+    });
+    game.addVillage('sensorario', new Village(
+        'sensorario',
+        'sensorario\'s village'
+    ));
 })