+++ CHANGELOG.md
@@ -8,3 +8,6 @@
  - add building to the building tree
  - extract building requirements
  - only first item in building tree can have empty requirements
+ - create village when new user is added
+ - display number of villages in example html
+ - display number of fields created